+++ CHANGELOG.md
@@ -69,6 +69,8 @@ Enter the dimension by throwing an enderpearl at Bee Nest and exit it by going a
 - Blacklisting any Bumblezone Biomes with Resourceful Bees's biome blacklist will prevent the bee from spawning at all in Bumblezone's Dimension.
 
 - Added Charm support so now their candles spawn in Bee Dungeons and Spider Infested Bee Dungeons!
+  
+- Several Bumblezone Items are now used in Charm's, ReosurcefulBees's, and Buzzier Bee's BeeKeeper/Apiarist trade offers!
 
 - Carrier Bees's bees now cannot get Wrath of the Hive effect and gets properly angered too now when a mob gets the effect.